Refactor: Move DeviceExt to device_ext.py and rename base.py to base_pin_memory.py
